tools/rrnd_test/requirements.txt
pyzmq
nose
numpy
scipy
flake8
skyfield
pyephem
matplotlib
